store receivers after showing config screen
The config screen allows the user to set up specific panic responses in the
receiving app.  It also acts as a TOFU prompt for the receiving app to
store info about the sender of the panic trigger to make sure that it only
responds to panic triggers from trusted apps.