do not use returned value to set which receivers are trusted
The old logic put the packageName in the Intent, then used the packageName
from the returned Intent in onActivityResult().  That would let the
receiver stick anything it wanted there, and it would have been added to
the list.

Instead, this code remembers the packageName where the Intent was sent,
then uses that in onActivityResult().