handle ACTION_DISCONNECT events
The receiving app can send a disconnect at any time to tell the trigger app
that it no longer wants to be connected.  This handles that Intent silently
in the background, since it was most likely sent because the user removed
this app using the UI in the receiving app.

This also reduces the scope of "Set<String> receiverPackageNames" to only
onResume().